separate find logic per target os
