+++ ENIGMAsystem/SHELL/Universal_System/Extensions/ttf/find-osx.cpp
@@ -1,0 +1,9 @@
+#include "find.h"
+
+namespace enigma {
+
+std::string font_lookup(std::string name, bool bold, bool italic) {
+  return "";
+}
+
+}
