+++ ENIGMAsystem/SHELL/Universal_System/Extensions/ttf/find.h
@@ -1,0 +1,5 @@
+#include <string>
+
+namespace enigma {
+std::string font_lookup(std::string name, bool bold, bool italic);
+}
